Add files via upload
Week 2: Programs to Understand Structure & Unions.
Question 1:- Define a structure to represent a student. Keep track of you, ID, Name and CGPA. Write function namely input, display and edit to input details to a student variable, to display it and to edit it respectively. Write a main function with define a structure pointer to access variables. Invoke the functions in the main and observe the result.